Update index.md
If a person uses the app they made in the previous step, it won't be the same template. That caused me quite a bit of confusion. I think it makes sense to start with a new app, and explicitly point out that you're using a template (which isn't the default one and will therefore be different from the app they may have just made).