Update loading.html
Added example to show how to use loading the async way. Otherwise its hard to see (without looking into the code) that it can also be used in an asynchronous way (which is most likely the default).

I noticed this:  Use of the `destroy()` function on the loading ViewController causes undesired side effects. Details in this post:
https://medium.com/@pascal.brokmeier/beware-of-using-destroy-on-ionic-2-loading-8f81c1ffe634#.6qnt6wpwt