docs(): update sass docs
This helps because you can just filter what you need and then copy and paste it into your own app.variables.scss. then you just change the values to your gusto and thanks to live injection of sass we see the changes immediately. it just saves the manual addition of the ": " between the variable names and the values after copying.