Clarificate Deploying requirements
The requirement of having Android Studio installed is optional if we take into account that we can also use Android SDK's command line tools. You can install, update, and uninstall packages with one of those tools (`sdkmanager`). This is also offered as an option in their site, and the download has a smaller file size.