CONTRIBUTING.KOR.md
Hi. Official I am a college student in Korea. I've translated the CONTRIBUTING.md file to make it easier for Korean contributors to access and contribute to the ionic web-site.
I'm still a beginner, so I have a lot. Please let me know if there is anything you need to fix.